refactor(app): app settings
