+++ app/src/organisms/Desktop/AdvancedSettings/__tests__/ShowHeaterShakerAttachmentModal.test.tsx
