+++ app/src/organisms/Desktop/AdvancedSettings/__tests__/U2EInformation.test.tsx
