+++ app/src/organisms/Desktop/AppSettings/__tests__/ConnectRobotSlideout.test.tsx
