+++ app/src/pages/Desktop/AppSettings/AdvancedSettings.tsx
@@ -11,7 +11,7 @@ import {
   U2EInformation,
   UpdatedChannel,
   AdditionalCustomLabwareSourceFolder,
-} from '/app/organisms/AdvancedSettings'
+} from '/app/organisms/Desktop/AdvancedSettings'
 
 export function AdvancedSettings(): JSX.Element {
   return (
