+++ app/src/pages/Desktop/AppSettings/GeneralSettings.tsx
@@ -41,8 +41,8 @@ import {
   ANALYTICS_APP_UPDATE_NOTIFICATIONS_TOGGLED,
 } from '/app/redux/analytics'
 import { UpdateAppModal } from '/app/organisms/Desktop/UpdateAppModal'
-import { PreviousVersionModal } from '/app/organisms/AppSettings/PreviousVersionModal'
-import { ConnectRobotSlideout } from '/app/organisms/AppSettings/ConnectRobotSlideout'
+import { PreviousVersionModal } from '/app/organisms/Desktop/AppSettings/PreviousVersionModal'
+import { ConnectRobotSlideout } from '/app/organisms/Desktop/AppSettings/ConnectRobotSlideout'
 import { getTopPortalEl } from '/app/App/portal'
 
 import type { Dispatch, State } from '/app/redux/types'
